+++ server/test-b64.js
@@ -1,4 +1,0 @@
-const buf = Buffer.from("");
-console.log(buf.length);
-console.log(buf.toString('base64') === "");
-console.log(!buf.toString('base64'));